Fix: When using --image with exec codex hangs, and fix image detection in exec prompt (#2795)
