+++ codex-rs/common/Cargo.toml
@@ -12,9 +12,12 @@ codex-core = { path = "../core" }
 codex-protocol = { path = "../protocol" }
 serde = { version = "1", optional = true }
 toml = { version = "0.9", optional = true }
+url = { version = "2", optional = true }
+shlex = { version = "1.3.0", optional = true }
 
 [features]
 # Separate feature so that `clap` is not a mandatory dependency.
 cli = ["clap", "serde", "toml"]
 elapsed = []
 sandbox_summary = []
+path_utils = ["url", "shlex"]
