+++ codex-rs/common/src/lib.rs
@@ -34,3 +34,6 @@ pub mod model_presets;
 // Shared approval presets (AskForApproval + Sandbox) used by TUI and MCP server
 // Not to be confused with AskForApproval, which we should probably rename to EscalationPolicy.
 pub mod approval_presets;
+// Shared path and image utilities used by TUI and exec
+#[cfg(feature = "path_utils")]
+pub mod path_utils;
