+++ codex-rs/exec/Cargo.toml
@@ -23,6 +23,7 @@ codex-common = { path = "../common", features = [
     "cli",
     "elapsed",
     "sandbox_summary",
+    "path_utils",
 ] }
 codex-core = { path = "../core" }
 codex-login = { path = "../login" }
@@ -31,6 +32,11 @@ codex-protocol = { path = "../protocol" }
 owo-colors = "4.2.0"
 serde_json = "1"
 shlex = "1.3.0"
+image = { version = "^0.25.6", default-features = false, features = [
+    "jpeg",
+    "png",
+] }
+url = "2"
 tokio = { version = "1", features = [
     "io-std",
     "macros",
